+++ studio/commands/__init__.py
@@ -3,6 +3,8 @@
 from studio.commands.command import Command
 from studio.commands.command_manager import CommandManager
 from studio.commands.delete_piece_command import DeletePieceCommand
+from studio.commands.import_boards_command import ImportBoardsCommand
+from studio.commands.import_pieces_command import ImportPiecesCommand
 from studio.commands.move_piece_command import MovePieceCommand
 from studio.commands.rotate_piece_command import RotatePieceCommand
 
@@ -10,6 +12,8 @@ __all__ = [
     "Command",
     "CommandManager",
     "DeletePieceCommand",
+    "ImportBoardsCommand",
+    "ImportPiecesCommand",
     "MovePieceCommand",
     "RotatePieceCommand",
 ]